# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Librarr API",
    "description": "Self-hosted book, audiobook, and manga search and download manager — the *arr for books. The HTTP/JSON API powers both the SPA and any external tooling (CLI scripts, AI agents, *arr-family integrations).",
    "version": "1.x",
    "license": {"name": "MIT", "identifier": "MIT"}
  },
  "servers": [
    {"url": "http://localhost:5050", "description": "Default local install"}
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key for programmatic access. Also accepted via ?apikey= query parameter."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "librarr_session",
        "description": "Set by POST /api/login. Used by the SPA."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {"type": "boolean", "enum": [false]},
          "error":   {"type": "string"}
        }
      },
      "SearchResult": {
        "type": "object",
        "description": "A single search hit. Fields are populated per driver — direct-download drivers omit seeders/leechers; torrent drivers usually omit MD5.",
        "properties": {
          "title":        {"type": "string"},
          "author":       {"type": "string"},
          "source":       {"type": "string", "description": "Driver identifier (e.g. annas, gutenberg, prowlarr)."},
          "size":         {"type": "integer", "format": "int64"},
          "size_human":   {"type": "string"},
          "seeders":      {"type": "integer"},
          "leechers":     {"type": "integer"},
          "magnet_url":   {"type": "string"},
          "download_url": {"type": "string", "format": "uri"},
          "md5":          {"type": "string"},
          "info_hash":    {"type": "string"},
          "url":          {"type": "string", "format": "uri"},
          "cover_url":    {"type": "string", "format": "uri"},
          "format":       {"type": "string", "description": "epub, pdf, mobi, mp3, m4b, cbz, etc."},
          "score":        {"type": "integer", "description": "0-100 confidence score."},
          "media_type":   {"type": "string", "enum": ["ebook", "audiobook", "manga"]}
        }
      },
      "WishlistItem": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "id":         {"type": "integer", "format": "int64"},
          "title":      {"type": "string"},
          "author":     {"type": "string"},
          "media_type": {"type": "string", "enum": ["ebook", "audiobook", "manga"]}
        }
      },
      "PaginatedList": {
        "type": "object",
        "description": "Standard list response shape used across /api/library, /api/activity, etc.",
        "properties": {
          "items":  {"type": "array", "items": {"type": "object"}},
          "total":  {"type": "integer"},
          "limit":  {"type": "integer"},
          "offset": {"type": "integer"}
        }
      }
    }
  },
  "security": [{"apiKey": []}, {"session": []}],
  "paths": {
...
    "/api/search": {
      "get": {
        "summary": "Search ebooks across all configured sources",
        "parameters": [
          {"name": "q",      "in": "query", "required": true,  "schema": {"type": "string"}, "description": "Search query."},
          {"name": "author", "in": "query", "required": false, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results":        {"type": "array", "items": {"$ref": "#/components/schemas/SearchResult"}},
                    "search_time_ms": {"type": "integer"},
                    "sources":        {"type": "array", "items": {"type": "object"}}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/search/audiobooks": {
      "get": {
        "summary": "Search audiobooks",
        "parameters": [
          {"name": "q",      "in": "query", "required": true,  "schema": {"type": "string"}},
          {"name": "author", "in": "query", "required": false, "schema": {"type": "string"}}
        ]
      }
    },
    "/api/search/manga": {
      "get": {
        "summary": "Search manga",
        "parameters": [{"name": "q", "in": "query", "required": true, "schema": {"type": "string"}}]
      }
    },
...
    "/api/library": {
      "get": {
        "summary": "List ebooks in the library",
        "parameters": [
          {"name": "limit",  "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 500, "default": 50}},
          {"name": "offset", "in": "query", "schema": {"type": "integer", "minimum": 0, "default": 0}},
          {"name": "type",   "in": "query", "schema": {"type": "string", "enum": ["ebook", "audiobook", "manga"]}},
          {"name": "tag",    "in": "query", "schema": {"type": "string"}, "description": "Filter by tag name."}
        ],
        "responses": {
          "200": {
            "description": "Paginated library items",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedList"}}}
          }
        }
      }
    },
...
    "/api/library/book/{id}":      {"delete": {"summary": "Remove an ebook from the library", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}]}},
    "/api/library/audiobook/{id}": {"delete": {"summary": "Remove an audiobook", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}]}},
...
    "/api/wishlist": {
      "get":  {"summary": "List wishlist items"},
      "post": {
        "summary": "Add a wishlist item",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WishlistItem"}}}
        },
        "responses": {
          "201": {"description": "Created"},
          "400": {"description": "Validation error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
    "/api/wishlist/{id}": {
      "delete": {
        "summary": "Delete a wishlist item",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}]
      }
    },
...
    "/torznab/api": {
      "get": {
        "summary": "Torznab indexer API",
        "description": "Lets Prowlarr / Readarr / other *arr apps use Librarr as a Torznab indexer.",
        "parameters": [
          {"name": "t",      "in": "query", "schema": {"type": "string", "enum": ["caps", "search", "book", "audio"]}},
          {"name": "q",      "in": "query", "schema": {"type": "string"}},
          {"name": "title",  "in": "query", "schema": {"type": "string"}},
          {"name": "author", "in": "query", "schema": {"type": "string"}},
          {"name": "apikey", "in": "query", "schema": {"type": "string"}, "description": "Required when TORZNAB_API_KEY is set."}
        ],
        "responses": {
          "200": {"description": "Torznab XML (RSS feed or caps document)", "content": {"application/xml": {}}}
        }
      }
    },
...
  }
}